+++ R/rarefyBioTIME_reference.R
@@ -14,8 +14,6 @@
 #' whenever a \code{NA}
 #' is found in the measure field(s), the whole sample is removed instead of the
 #' missing observations only.
-#' @param summarise (\code{logical}). \code{TRUE} by default. If \code{FALSE},
-#' the function returns abundance and/or biomass summed at the
 #'
 #' @returns Returns a single long form \code{data.frame} containing the total currency
 #' or currencies of interest (sum) for each species in each year within each
@@ -24,10 +22,7 @@
 #'
 #' @details
 #' Sample-based rarefaction prevents temporal variation in sampling effort from
-#' affecting diversity estimates (see Gotelli N.J., Colwell R.K. 2001 Quantifying
-#' biodiversity: procedures and pitfalls in the measurement and comparison of species
-#' richness. Ecology Letters 4(4), 379-391) by selecting an equal number of samples
-#' across all years in a time series.
+#' affecting diversity estimates (see Gotelli N.J., Colwell R.K. 2001 Quantifying biodiversity: procedures and pitfalls in the measurement and comparison of species richness. Ecology Letters 4(4), 379-391) by selecting an equal number of samples across all years in a time series.
 #' \code{resampling} counts the number of unique samples taken in each year (sampling effort),
 #' identifies the minimum number of samples across all years, and then uses this minimum to
 #' randomly resample each year down to that number. Thus, standardising the
@@ -56,21 +51,13 @@
 #'   library(BioTIMEr)
 #'   set.seed(42)
 #'   x <- gridding(BTsubset_meta, BTsubset_data)
-#'   resampling(x, measure = "BIOMASS", summarise = TRUE)
+#'   resampling(x, measure = "BIOMASS")
 #'   resampling(x, measure = "ABUNDANCE")
 #'   resampling(x, measure = c("ABUNDANCE","BIOMASS"))
-#'   # Without summarising the species abundances are summed at the SAMPLE_DESC level
-#'   resampling(x, measure = "BIOMASS", summarise = FALSE, conservative = FALSE)
 #' }
 #'
 
-resampling <- function(
-  x,
-  measure,
-  resamps = 1L,
-  conservative = FALSE,
-  summarise = TRUE
-) {
+resampling_ref <- function(x, measure, resamps = 1L, conservative = FALSE) {
   if (inherits(x, "data.table")) {
     warning("x was converted to a data.frame")
     x <- as.data.frame(x)
@@ -90,12 +77,6 @@ resampling <- function(
   checkmate::assert_integer(x = x$YEAR, lower = 1300L, null.ok = FALSE)
   checkmate::assert_logical(
     x = conservative,
-    len = 1L,
-    null.ok = FALSE,
-    any.missing = FALSE
-  )
-  checkmate::assert_logical(
-    x = summarise,
     len = 1L,
     null.ok = FALSE,
     any.missing = FALSE
@@ -151,11 +132,10 @@ resampling <- function(
   }
 
   if (
-    tapply(x$YEAR, x$STUDY_ID, function(y) {
-      dplyr::n_distinct(y) == 1L
-    }) |>
+    tapply(x$YEAR, x$STUDY_ID, function(y) length(unique(y)) == 1L) |>
       any()
   ) {
+    warning("Some 1-year-long studies were removed.")
     x <- dplyr::anti_join(
       x = x,
       y = x |>
@@ -166,28 +146,24 @@ resampling <- function(
         dplyr::filter(.data$count == 1L),
       by = dplyr::join_by("STUDY_ID")
     )
-    warning("Some 1-year-long studies were removed.")
   }
 
   rfIDs <- unique(x$assemblageID)
   TSrf <- sapply(
     X = rfIDs,
     FUN = function(i) {
-      x[x$assemblageID == i, ] |>
-        rarefysamples(
-          measure = measure,
-          resamps = resamps,
-          summarise = summarise
-        )
+      temp_data <- x[x$assemblageID == i, ]
+      rarefysamples_ref(x = temp_data, measure = measure, resamps = resamps)
     },
     USE.NAMES = TRUE,
     simplify = FALSE
   )
 
   return({
-    dplyr::bind_rows(TSrf, .id = "rfID") |>
+    dplyr::bind_rows(TSrf) |>
+      dplyr::mutate(rfID = rep(rfIDs, times = sapply(TSrf, nrow))) |>
       tidyr::separate(
-        col = "rfID",
+        "rfID",
         into = c("STUDY_ID", "cell"),
         sep = "_",
         remove = FALSE
@@ -199,7 +175,6 @@ resampling <- function(
         "STUDY_ID",
         "YEAR",
         "Species",
-        dplyr::any_of(c("SAMPLE_DESC", "minsamp")),
         dplyr::all_of(measure)
       )
   })
@@ -209,58 +184,43 @@ resampling <- function(
 #' Rarefy BioTIME data
 #' Applies sample-based rarefaction to standardise the number of samples per year
 #'    within a cell-level time series.
-#' @inheritParams resampling
+#' @inheritParams resampling_ref
 #' @returns Returns a single long form data frame containing the total currency
 #'    of interest (sum) for each species in each year.
 #' @keywords internal
 
-rarefysamples <- function(x, measure, resamps, summarise) {
+rarefysamples_ref <- function(x, measure, resamps) {
   # Computing minimal effort per year in this assemblageID
-  minsample <- tapply(x$SAMPLE_DESC, x$YEAR, function(x) {
-    dplyr::n_distinct(x)
-  }) |>
-    min()
+  minsample <- min(tapply(x$SAMPLE_DESC, x$YEAR, function(x) length(unique(x))))
 
   rareftab_list <- lapply(
     # beginning loop on repetitions
     X = seq_len(resamps),
     FUN = function(i) {
-      selected_indices <- lapply(
+      selected_indices <- unlist(lapply(
         # beginning sub loop on years
         X = unique(x$YEAR),
         FUN = function(y) {
-          sam <- x$SAMPLE_DESC[x$YEAR == y] |>
-            unique() |>
-            sample(minsample, replace = FALSE)
+          samps <- unique(x$SAMPLE_DESC[x$YEAR == y])
+          sam <- sample(samps, minsample, replace = FALSE)
           return(which(x$SAMPLE_DESC %in% sam & x$YEAR == y))
         }
-      ) |>
-        unlist() # end of loop on years
+      )) # end of loop on years
 
-      if (summarise) {
-        raref <- stats::aggregate(
-          x = x[selected_indices, measure, drop = FALSE],
-          by = list(
-            YEAR = x$YEAR[selected_indices],
-            Species = x$Species[selected_indices]
-          ),
-          FUN = sum
-        )
-      } else {
-        # Does it really make a difference?
-        raref <- stats::aggregate(
-          x = x[selected_indices, measure, drop = FALSE],
-          by = list(
-            SAMPLE_DESC = x$SAMPLE_DESC[selected_indices],
-            YEAR = x$YEAR[selected_indices],
-            Species = x$Species[selected_indices]
-          ),
-          FUN = sum
-        )
-        raref$minsamp = minsample
-      }
+      tYear <- x[selected_indices, "YEAR"]
+      tSpecies <- x[selected_indices, "Species"]
+      tcurrency <- x[selected_indices, measure, drop = FALSE]
+
+      raref <- stats::aggregate(
+        x = tcurrency,
+        by = list(YEAR = tYear, Species = tSpecies),
+        FUN = sum
+      )
+      raref$resamp <- i
       return(raref)
     }
   ) # end of loop on repetitions
-  return(dplyr::bind_rows(rareftab_list, .id = "resamp"))
+
+  rareftab <- dplyr::bind_rows(rareftab_list)
+  return(rareftab)
 } # end of function